done lab3 (REST)

+++ src/main/java/com/example/esa_lab_2/service/PaperService.java
@@ -1,14 +1,17 @@
 package com.example.esa_lab_2.service;
 
 
+import com.example.esa_lab_2.dto.PaperCreationDTO;
 import com.example.esa_lab_2.model.Paper;
 import com.example.esa_lab_2.model.Participant;
 import com.example.esa_lab_2.repository.PaperRepository;
 import com.example.esa_lab_2.repository.ParticipantRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -25,11 +28,34 @@ public class PaperService {
         return paperRepository.findByAuthor(author);
     }
 
+    public Optional<Paper> getById(Long id) {
+        return paperRepository.findById(id);
+    }
+
     public void create(Paper newPaper) {
         paperRepository.save(newPaper);
     }
 
+    public void create(PaperCreationDTO dto) {
+        Participant author = participantRepository.findByName(dto.getAuthor());
+
+        if (author != null) {
+            Paper paper = Paper.builder()
+                    .title(dto.getTitle())
+                    .year(dto.getYear())
+                    .author(author)
+                    .build();
+            paperRepository.save(paper);
+        }
+    }
+
     public void deleteById(Long id) {
-        paperRepository.deleteById(id);
+        Optional<Paper> paper = paperRepository.findById(id);
+        if (paper.isPresent()) {
+            paperRepository.deleteById(id);
+        }
+        else {
+            throw new EntityNotFoundException();
+        }
     }
 }
