+++ src/main/java/com/example/esa_lab_2/service/ParticipantService.java
@@ -1,7 +1,9 @@
 package com.example.esa_lab_2.service;
 
+import com.example.esa_lab_2.dto.ParticipantCreationDTO;
 import com.example.esa_lab_2.model.Participant;
 import com.example.esa_lab_2.repository.ParticipantRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -29,7 +31,20 @@ public class ParticipantService {
         participantRepository.save(newParticipant);
     }
 
+    public void create(ParticipantCreationDTO dto) {
+        Participant participant = Participant.builder()
+                .name(dto.getName())
+                .age(dto.getAge())
+                .academicDegree(dto.getAcademicDegree())
+                .build();
+        participantRepository.save(participant);
+    }
+
     public void deleteById(Long id) {
-        participantRepository.deleteById(id);
+        Optional<Participant> participant = participantRepository.findById(id);
+        if (participant.isPresent())
+            participantRepository.deleteById(id);
+        else
+            throw new EntityNotFoundException();
     }
 }
